pyproject.toml
[project]
name = "openhands-gepa-sre"
version = "0.1.0"
description = "Self-Optimizing SRE Agent demo using OpenHands SDK + GEPA-style optimization"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "flask==3.1.0",
  "dspy-ai>=2.5.0",
  "openhands-sdk>=1.0.0",
  "python-dotenv>=1.0.0",
  "openhands-tools>=1.11.5",
]

[tool.uv]
package = false
